move first-time-setup components
