+++ src/common/components/login/hidive.vue
@@ -100,7 +100,7 @@ export default class LoginHd extends Vue {
 </script>
 
 <style scoped lang="scss">
-@import '../../colors';
+@import '../../../colors';
 
 .login-hidive {
   position: relative;
