+++ src/views/settings/settings.vue
@@ -252,8 +252,8 @@ import Checkbox from '@/components/Checkbox.vue'
 import CButton from '@/components/CButton.vue'
 import Icon from '@/components/Icon.vue'
 import Dropdown, { DropdownItem } from '@/components/Form/Dropdown.vue'
-import LoginHD from '@/components/FirstTimeSetup/LoginHD.vue'
-import LoginCR from '@/components/FirstTimeSetup/LoginCR.vue'
+import LoginHD from '@/common/components/login/hidive.vue'
+import LoginCR from '@/common/components/login/crunchyroll.vue'
 
 import { Crunchyroll } from '@/lib/crunchyroll'
 import { getIsUpdateAvailable } from '@/state/app'
